pyproject.toml
[project]
name = "auraassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.174.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
